+++ common/src/main/java/earth/terrarium/chipped/common/blocks/DirectionBlock.java
@@ -1,0 +1,34 @@
+package earth.terrarium.chipped.common.blocks;
+
+import com.mojang.serialization.MapCodec;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.HorizontalDirectionalBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+
+public class DirectionBlock extends HorizontalDirectionalBlock {
+    public static final MapCodec<? extends HorizontalDirectionalBlock> CODEC = simpleCodec(DirectionBlock::new);
+
+    public DirectionBlock(Properties properties) {
+        super(properties);
+        registerDefaultState(this.defaultBlockState()
+            .setValue(FACING, Direction.NORTH));
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(FACING);
+    }
+
+    @Override
+    protected MapCodec<? extends HorizontalDirectionalBlock> codec() {
+        return CODEC;
+    }
+
+    @Override
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        return this.defaultBlockState().setValue(FACING, context.getHorizontalDirection().getOpposite());
+    }
+}